requirements.txt
Flask
elasticsearch==8.18.0
redis
python-dotenv